feat: 新增惠民卡联盟功能及多个门户支持
- 新增惠民卡联盟功能，包含京东、拼多多商品展示与推广
- 添加商家、技工、服务商三个门户的初始代码和配置
- 更新数据库模型和迁移脚本，支持新功能
- 优化.gitignore和.env配置
- 修复管理后台登录密码校验逻辑
- 移除不再使用的直播和商品相关代码
- 新增多个实用脚本和API文档更新

+++ backend/src/models/MarketCartItem.js
@@ -5,9 +5,9 @@ module.exports = (sequelize, DataTypes) => {
     static associate(models) {}
   }
   MarketCartItem.init({
-    user_id: { type: DataTypes.BIGINT, allowNull: false },
-    shop_id: { type: DataTypes.BIGINT, allowNull: false },
-    goods_id: { type: DataTypes.BIGINT, allowNull: false },
+    user_id: { type: DataTypes.INTEGER, allowNull: false },
+    shop_id: { type: DataTypes.INTEGER, allowNull: false },
+    goods_id: { type: DataTypes.INTEGER, allowNull: false },
     quantity: { type: DataTypes.INTEGER, allowNull: false, defaultValue: 1 },
     checked: { type: DataTypes.TINYINT, allowNull: false, defaultValue: 1 }
   }, {
