+++ backend/src/models/MerchantAccount.js
@@ -7,7 +7,7 @@ module.exports = (sequelize, DataTypes) => {
     }
   }
   MerchantAccount.init({
-    shop_id: { type: DataTypes.BIGINT, allowNull: false },
+    shop_id: { type: DataTypes.INTEGER, allowNull: false },
     username: { type: DataTypes.STRING(80), allowNull: false },
     password_hash: { type: DataTypes.STRING(255), allowNull: false },
     role: { type: DataTypes.ENUM('owner', 'manager', 'operator'), allowNull: false, defaultValue: 'operator' },
